Crate pallet_collection

Collection Module

Config

Call

Collection is used to represent a series of NonFungible or Fungible Tokens, which can also be understood as a folder. It is one of the basic modules.

Interface

Dispatchable Functions

create_collection - Create a collection to represent NFT/FT.

Re-exports

pub use $crate::traits::GetPalletVersion as _;

Structs

CollectionInfo
Details of a collection.

Collections
The set of collection.

Module
The module declaration.

Nonce
Increment number used to create collection_id.

Enums

Call
Dispatchable calls.

Error
Errors for this module.

RawEvent
Events for this module.

TokenType
Used to indicate the type of tokens in the Collection.

Traits

CollectionInterface

Config

Type Definitions

Event
RawEvent specialized for the configuration Config